+++ Engine/Source/Engine/Graphics/SwapChain.cpp
@@ -127,8 +127,8 @@ namespace Engine
 		                                                  renderTargetViews.data(),
 		                                                  &m_MainFramebuffer->GetDepthStencil());
 		D3D11_VIEWPORT vp;
-		vp.Width    = width;
-		vp.Height   = height;
+		vp.Width    = (FLOAT)width;
+		vp.Height   = (FLOAT)height;
 		vp.MinDepth = 0.0f;
 		vp.MaxDepth = 1.0f;
 		vp.TopLeftX = 0;
